authRefactor: google and facebook working.
